chore(contracts): factor out function into `utils` module
